[FIX] Change the type of cflags to match the struct.

+++ exec.c
@@ -675,7 +675,7 @@ static void build_page_bitmap(PageDesc *p)
     }
 }
 
-TranslationBlock *tb_gen_code(CPUState *env, target_ulong pc, target_ulong cs_base, int flags, int cflags)
+TranslationBlock *tb_gen_code(CPUState *env, target_ulong pc, target_ulong cs_base, int flags, uint16_t cflags)
 {
     TranslationBlock *tb;
     uint8_t *tc_ptr;
